+++ database/migrations/2021_08_11_000000_create_ledger_logs_table.php
@@ -16,9 +16,12 @@ class CreateLedgerLogsTable extends Migration
         Schema::create('ledger_logs', function (Blueprint $table) {
             $table->id();
             $table->string('action');
-            $table->integer('user_id');
+            $table->text('details')->nullable();
+            $table->integer('user_id')->nullable();
             $table->integer('loggable_id');
             $table->string('loggable_type');
+            $table->string('ip')->nullable();
+            $table->string('user_agent')->nullable();
             $table->timestamps();
         });
     }