+++ src/Models/LedgerLog.php
@@ -6,21 +6,45 @@ use Illuminate\Database\Eloquent\Model;
 
 class LedgerLog extends Model
 {
+    const ACTION = 'action';
+    const DETAILS = 'details';
+    const USER_ID = 'user_id';
+    const LOGGABLE_ID = 'loggable_id';
+    const LOGGABLE_TYPE = 'loggable_type';
+    const IP = 'ip';
+    const USER_AGENT = 'user_agent';
+
     /**
      * The attributes that are mass assignable.
      *
      * @var array
      */
     protected $fillable = [
-        'action',
-        'user_id',
-        'loggable_id',
-        'loggable_type',
+        self::ACTION,
+        self::DETAILS,
+        self::USER_ID,
+        self::LOGGABLE_ID,
+        self::LOGGABLE_TYPE,
+        self::IP,
+        self::USER_AGENT,
     ];
 
 
     public function loggable()
     {
         return $this->morphTo();
+    }
+    
+    public function getDetailsAttribute($value)
+    {
+        return json_decode($value, true) ?? $value;
+    }
+
+    /**
+     * Registers a relationship with the user responsible for the action performed
+     */
+    public function user()
+    {
+        return $this->belongsTo(config('ledger.user'), config('ledger.user_primary_key'));
     }
 }